finish up, change task kinda
