+++ frontend/src/pages/home.tsx
@@ -16,11 +16,11 @@ export default function HomePage() {
   const [user, setUser] = React.useState(BlankUser);
   const dispatch = useAppDispatch();
 
-  const currentUser = useAppSelector((state) => state.user.currentUser);
+  const current = useAppSelector((state) => state.user.current);
 
   React.useEffect(() => {
     const fetchUser = async () => {
-      setUser(currentUser || BlankUser);
+      setUser(current || BlankUser);
     };
     fetchUser();
   }, []);
