+++ frontend/src/pages/landing.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Typography } from '@mui/material';
+
+export default function LandingPage() {
+  return (
+    <div>
+      <Typography variant="h1">LANDING PAGE</Typography>
+    </div>
+  );
+}
